destiny_clock/requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
google-genai>=1.31.0
